requirements2.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
cssselect==1.2.0
fake-useragent==1.1.3
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
lxml==4.9.2
mpire==2.6.0
multidict==6.0.4
mysql-connector-python==8.0.33
numpy==1.24.2
parse==1.19.0
protobuf==3.20.3
pyee==8.2.2
Pygments==2.14.0
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.14
w3lib==2.1.1
websockets==10.4
yarl==1.8.2
zipp==3.11.0
